backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.36
httpx==0.25.2
python-dotenv==1.0.0
apscheduler==3.10.4
psycopg2-binary==2.9.9
pytz==2024.1
transformers==4.40.0
torch==2.2.2
vaderSentiment==3.3.2